Move diagnostics to a separate library
This will make it possible to share it across crates.
