+++ Cargo.toml
@@ -10,3 +10,6 @@ resolver = "2"
 [workspace.package]
 edition = "2024"
 license = "MIT OR Apache-2.0"
+
+[workspace.dependencies]
+diagnostics = { path = "tools/diagnostics" }
